backend: Work around bug in wait_regexp timeout
As it is we do have a timeout for the poll() on the serial console, if
no message appears there for a defined time we give up. This
unfortunatelly does not work if we manage to lock a kernel in a way that
produces messages every minute or so. So this commit applies the timeout
over the whole loop that reads lines from the console and exits if we
fail to have a match over that period.

Signed-off-by: Cyril Hrubis <metan@ucw.cz>