Fix regression in --exclude and --include switches
This patch is follow up for 13ff090 and c0f37f2 which were trying to fix
support for lists in exclude and include but at the same time broke
support for regex. Current version keep support for lists and fixing
regex string support for exclude and include.

Tested on these (TODO: add a CI test for it):
--exclude="access0[13]"
--exclude="access01 access03"
--exclude="access01, access03"
--exclude '_16\$|^(move_pages|access0[13]|msg(ctl|get|rcv|snd|stress)|sem(ctl|get|op|snd)|shm(at|ctl|dt|get)|fallocate06)'

Fixes: https://github.com/metan-ucw/runltp-ng/issues/32
Closes: https://github.com/metan-ucw/runltp-ng/pull/33

Suggested-by: Li Wang <liwang@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Anton Smorodskyi <asmorodskyi@suse.com>