CI: Remove Ubuntu releases
Groovy and Xenial are EOL now. impish does not have linux-headers
package. But update releases does not help because Focal (LTS, it'd be a
candidate to test) and Hirsute have problems to install
linux-headers-5.11.0-1021-azure (some strange timeout).

This means that 6999232 ("install_pkg: Fix kernel headers for Ubuntu")
was tested only locally.

Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>